Proposal: Add NEWS file generator
This is an initial idea for creating a NEWS file from annotated tag
commit messages.

The format isn't quite right (no newline separators between versions)
and I still wonder how to put that information of the tag yet to be
created ontop of the NEWS file because obviously "git tag -l" won't show
it.

Affects: #50

Signed-off-by: Kai Krakow <kai@kaishome.de>